Proposal: separate template and config
The existence of a notebook template is a very nice thing, since we can
define common blocks of code that we always rely on...

However the current implementation have a huge chunk right in the
beginning of the file with all the boilerplate code (e.g. configuring
jupyter extensions and the way things are displayed in the notebook).

I wonder if using a single file to centralise all this "boilerplate
jupyter configuration" makes sense... The idea is not to replace the
template, but rather get unimportant things (like configuring how many
columns pandas will display), out of the way.

This way a person can have a single line on top of the file:

```ipynb
%run ./config.ipynb
```

and start writing some text for good'old literal programming, and the
readers will not be distracted by the usual chores...

This approach also allow users to change things in one place (e.g.
font size for the graphs) and have it automatically changing in all the
notebooks, which can be quite neat.